Extract Genebodies

Analagous to GenomicRanges::promoters, this function returns ranges that start and end downstream of the TSS. When fix = "start", the function behaves differently depending on the sign of the end parameter. Currently, there's no way to set the ranges to continue a fixed distance past the original end sites when fix = "start".

genebodies(
  genelist,
  start = 300,
  end = -300,
  fix_start = "start",
  fix_end = "end",
  min_window = 500
)

Arguments

genelist
A GRanges object containing genes of interest.

start
When fix = "start", the distance downstream of the TSS where the new ranges should begin.

end
Where the ranges should end. When end = 0, the returned ranges keep the original end sites. When end < 0, the new ranges will end abs(end) number of bases before the original end site. When end > 0 and fix = "start", the new ends are fixed to end number of bases from the original start site.

min.window
The minimum size of a returned genebody length, after accounting for start and end parameters.

fix
If set to "end", function will return ranges centered around the end of the ranges. Negative values for start will begin the output ranges a fixed distance before the end of the input ranges; positive values will begin the ranges after the ends of the input ranges. The behavior of end is the same as for start, and errors will be returned if end < start.

Value

A GRanges object that may be shorter than genelist due to loss of short ranges.