Get signal counts at each position within regions of interest

Generate a matrix containing a row for each region of interest, and columns for each position (each base if binsize = 1) within each region.

getCountsByPositions(
  dataset.gr,
  regions.gr,
  binsize = 1,
  FUN = sum,
  simplify.multi.widths = c("list", "pad 0", "pad NA"),
  field = "score",
  ncores = detectCores()
)

Arguments

dataset.gr
A GRanges object in which signal is contained in metadata (typically in the "score" field).

regions.gr
A GRanges object containing all the regions of interest.

binsize
Size of bins (in bp) to use for counting within each range of regions.gr. Note that counts will not be length-normalized.

FUN
If binsize > 1, the function used to aggregate the signal within each bin. By default, the signal is summed, but any function operating on a numeric vector can be used.

simplify.multi.widths
A string indicating the output format if the ranges in regions.gr have variable widths. Default = "list". See details below.

field
The metadata field of dataset.gr to be counted. If length(field) > 1, the output is a list whose elements contain the output for generated each field.

ncores
Multiple cores can only be used if length(field) > 1.

Details

If the widths of all ranges in regions.gr are equal, a matrix is returned containing a row for each range in regions.gr, and a column for each bin. For input regions.gr with varying widths, setting simplify.multi.widths = "list" will output a list of variable-length vectors, with each vector corresponding to an input region. If simplify.multi.widths = "pad 0" or "pad NA", the output is a matrix containing a row for each range in regions.gr, and a column for each position in each range. The number of columns is determined by the largest range in regions.gr, and columns corresponding to positions outside of each range are either set to 0 or NA, depending on the argument.

See also

Examples

data("PROseq") # load included PROseq data
data("txs_dm6_chr4") # load included transcripts

#--------------------------------------------------#
# counts from 0 to 50 bp after the TSS
#--------------------------------------------------#

txs_pr <- promoters(txs_dm6_chr4, 0, 50) # first 50 bases
countsmat <- getCountsByPositions(PROseq, txs_pr)
countsmat[1:6, 40:50] # show only 40-50 bp after TSS

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#> [1,]    0    0    0    0    0    0    0    0    0     0     0
#> [2,]    0    0    0    0    0    0    0    0    0     0     0
#> [3,]    0    0    0    0    0    0    0    0    0     0     0
#> [4,]    0    0    0    0    0    0    0    0    0     0     0
#> [5,]    0    0    0    0    0    0    0    0    0     0     0
#> [6,]    0    0    0    1    0    0    0    0    0     0     0

#--------------------------------------------------#
# redo with 10 bp bins from 0 to 100
#--------------------------------------------------#

txs_pr <- promoters(txs_dm6_chr4, 0, 100)
countsmat <- getCountsByPositions(PROseq, txs_pr, binsize = 10)
countsmat[1:6, ]

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    0    0    0    0    0    0    0    0    0     0
#> [2,]    2    1    7    1    0    0    0    1    4     0
#> [3,]    0    0    1    0    0    0    0    1    0     0
#> [4,]    0    0    0    0    0    0    0    0    0     0
#> [5,]    0    0    0    0    0    0    0    0    0     0
#> [6,]    0    0    0    0    1   12   20  160    4    12

#--------------------------------------------------#
# same as the above, but with the average signal in each bin
#--------------------------------------------------#

countsmat <- getCountsByPositions(PROseq, txs_pr, binsize = 10, FUN = mean)
countsmat[1:6, ]

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]  0.0  0.0  0.0  0.0  0.0  0.0    0  0.0  0.0   0.0
#> [2,]  0.2  0.1  0.7  0.1  0.0  0.0    0  0.1  0.4   0.0
#> [3,]  0.0  0.0  0.1  0.0  0.0  0.0    0  0.1  0.0   0.0
#> [4,]  0.0  0.0  0.0  0.0  0.0  0.0    0  0.0  0.0   0.0
#> [5,]  0.0  0.0  0.0  0.0  0.0  0.0    0  0.0  0.0   0.0
#> [6,]  0.0  0.0  0.0  0.0  0.1  1.2    2 16.0  0.4   1.2

#--------------------------------------------------#
# standard deviation of signal in each bin
#--------------------------------------------------#

countsmat <- getCountsByPositions(PROseq, txs_pr, binsize = 10, FUN = sd)
round(countsmat[1:6, ], 2)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7]  [,8] [,9] [,10]
#> [1,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.0  0.00
#> [2,] 0.42 0.32 1.06 0.32 0.00 0.00 0.00  0.32  0.7  0.00
#> [3,] 0.00 0.00 0.32 0.00 0.00 0.00 0.00  0.32  0.0  0.00
#> [4,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.0  0.00
#> [5,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.0  0.00
#> [6,] 0.00 0.00 0.00 0.00 0.32 3.79 4.06 38.00  0.7  0.79