Find sites with max signal in regions of interest

For each signal-containing region of interest, find the single site with the most signal. Sites can be found at base-pair resolution, or defined for larger bins.

getMaxPositionsBySignal(
  regions.gr,
  dataset.gr,
  binsize = 1,
  bin.centers = FALSE,
  field = "score",
  keep.score = FALSE
)

Arguments

regions.gr
A GRanges object containing regions of interest.

dataset.gr
A GRanges object in which signal is contained in metadata (typically in the "score" field).

binsize
The size of bin in which to calculate signal scores.

bin.centers
Logical indicating if the centers of bins are returned, as opposed to the entire bin. If TRUE,

field
The metadata field of dataset.gr to be counted.

keep.score
Logical indicating if the signal value at the max site should be reported. If set to TRUE, the values are kept as a new metadata column in regions.gr.

Value

Output is a GRanges object with regions.gr metadata, but each range only contains the site within each regions.gr range that had the most signal. If binsize > 1, the entire bin is returned, unless bin.centers = TRUE, in which case a single-base site is returned. The site is set to the center of the bin, and if the binsize is even, the site is rounded to be closer to the beginning of the range.

If keep.score = TRUE, the output will also contain metadata for the signal at the max site. The output is not necessarily same length as regions.gr, as regions without signal are not returned. If no regions have signal (e.g. as could happen if running this function on a single region), the function will return an empty GRanges object with intact metadata columns.

See also

Examples

data("PROseq") # load included PROseq data
data("txs_dm6_chr4") # load included transcripts

#--------------------------------------------------#
# first 50 bases of transcripts
#--------------------------------------------------#

pr <- promoters(txs_dm6_chr4, 0, 50)
pr[1:3]

#> GRanges object with 3 ranges and 2 metadata columns:
#>       seqnames      ranges strand |     tx_name     gene_id
#>          <Rle>   <IRanges>  <Rle> | <character> <character>
#>   [1]     chr4     879-928      + | FBtr0346692 FBgn0267363
#>   [2]     chr4 42774-42823      + | FBtr0344900 FBgn0266617
#>   [3]     chr4 44774-44823      + | FBtr0340499 FBgn0265633
#>   -------
#>   seqinfo: 7 sequences from dm6 genome

#--------------------------------------------------#
# max sites
#--------------------------------------------------#

getMaxPositionsBySignal(pr[1:3], PROseq, keep.score = TRUE)

#> GRanges object with 2 ranges and 3 metadata columns:
#>       seqnames    ranges strand |     tx_name     gene_id MaxSiteScore
#>          <Rle> <IRanges>  <Rle> | <character> <character>    <numeric>
#>   [1]     chr4     42799      + | FBtr0344900 FBgn0266617            3
#>   [2]     chr4     44801      + | FBtr0340499 FBgn0265633            1
#>   -------
#>   seqinfo: 7 sequences from dm6 genome

#--------------------------------------------------#
# max sites in 5 bp bins
#--------------------------------------------------#

getMaxPositionsBySignal(pr[1:3], PROseq, binsize = 5, keep.score = TRUE)

#> GRanges object with 2 ranges and 3 metadata columns:
#>       seqnames      ranges strand |     tx_name     gene_id MaxSiteScore
#>          <Rle>   <IRanges>  <Rle> | <character> <character>    <numeric>
#>   [1]     chr4 42794-42798      + | FBtr0344900 FBgn0266617            4
#>   [2]     chr4 44799-44803      + | FBtr0340499 FBgn0265633            1
#>   -------
#>   seqinfo: 7 sequences from dm6 genome