Calculate pausing indices from user-supplied promoters & genebodies

Pausing index (PI) is calculated for each gene (within matched promoters.gr and genebodies.gr) as promoter-proximal (or pause region) signal counts divided by genebody signal counts. If length.normalize = TRUE (recommended), the signal counts within each range in promoters.gr and genebodies.gr are divided by their respective range widths (region lengths) before pausing indices are calculated.

getPausingIndices(
  dataset.gr,
  promoters.gr,
  genebodies.gr,
  field = "score",
  length.normalize = TRUE,
  remove.empty = FALSE,
  ncores = detectCores()
)

Arguments

dataset.gr
A GRanges object in which signal is contained in metadata (typically in the "score" field).

promoters.gr
A GRanges object containing promoter-proximal regions of interest.

genebodies.gr
A GRanges object containing genebody regions of interest.

field
The metadata field of dataset.gr to be counted. If length(field) > 1, a dataframe is returned containing the pausing indices for each region in each field.

length.normalize
A logical indicating if signal counts within regions of interest should be length normalized. The default is TRUE, which is recommended, especially if input regions don't all have the same width.

remove.empty
A logical indicating if genes without any signal in promoters.gr should be removed. No genes are filtered by default.

ncores
Multiple cores can only be used if length(field) > 1.

Value

A vector of length given by the length of the genelist (or possibly shorter if remove.empty = TRUE). If length(field) > 1, a dataframe is returned, containing a column for each field.

See also

Examples

data("PROseq") # load included PROseq data
data("txs_dm6_chr4") # load included transcripts

#--------------------------------------------------#
# Get promoter-proximal and genebody regions
#--------------------------------------------------#

# genebodies from +300 to 300 bp before the poly-A site
gb <- genebodies(txs_dm6_chr4, 300, -300, min.window = 400)

# get the transcripts that are large enough (>1kb in size)
txs <- subset(txs_dm6_chr4, tx_name %in% gb$tx_name)

# for the same transcripts, promoter-proximal region from 0 to +100
pr <- promoters(txs, 0, 100)

#--------------------------------------------------#
# Calculate pausing indices
#--------------------------------------------------#

pidx <- getPausingIndices(PROseq, pr, gb)

length(txs_dm6_chr4)

#> [1] 339

length(pidx)

#> [1] 318

head(pidx)

#> [1]  0.00000  3.50500  0.00000  0.00000 29.03505 28.15400

#--------------------------------------------------#
# Without length normalization
#--------------------------------------------------#

head( getPausingIndices(PROseq, pr, gb, length.normalize = FALSE) )

#> [1] 0.00000000 0.50000000 0.00000000 0.00000000 0.09219232 0.07261987

#--------------------------------------------------#
# Removing empty means the values no longer match the genelist
#--------------------------------------------------#

pidx_signal <- getPausingIndices(PROseq, pr, gb, remove.empty = TRUE)

length(pidx_signal)

#> [1] 240