Import PRO-seq (or similar) bigWig files

This function imports plus/minus pairs of bigWig files containing basepair-resolution data, e.g. PRO-seq or PRO-cap data.

import.PROseq(
  plus_file,
  minus_file,
  genome = NULL,
  keep.X = TRUE,
  keep.Y = TRUE,
  keep.M = FALSE,
  keep.nonstandard = FALSE
)

Arguments

genome
Optional string for UCSC reference genome, e.g. "hg38". If given, non-standard chromosomes are trimmed, and options for sex and mitochondrial chromosomes are applied.

keep.X
Logical indicating whether the X chromosome should be kept.

keep.Y
Logical indicating whether the Y chromosome should be kept.

keep.M
Logical indicating whether mitochondrial chromosomes should be kept.

plus_bw
Path of plus strand bigWig file.

minus_bw
Path of minus strand bigWig file.

Value

Imports a GRanges object containing base-pair resolution data, with the score metadata column indicating readcounts at each base. All ranges are of width = 1.