Import bam files

Import bam files

import_bam(
  file,
  mapq = 20,
  revcomp = FALSE,
  trim.to = c("whole", "5p", "3p", "center"),
  ignore.strand = FALSE,
  field = "score"
)

Arguments

file
Path of a bam file.

mapq
Filter reads by a minimum MapQ score. This is the correct way to filter multi-aligners.

revcomp
Logical indicating if aligned reads should be reverse-complemented.

trim.to
Option for selecting specific bases from the reads, applied after the revcomp option. By default, the entire read is maintained. Other options are to take only the 5' base, only the 3' base, or the only the center base of the read.

ignore.strand
Logical indicating if the strand information should be discarded. If TRUE, strand information is discarded after revcomp and trim.to options are applied.

field
Metadata field name to use for readcounts, usually "score". If set to NULL, identical reads (or identical positions if trim.to options applied) are not combined, and the length of the output GRanges will be equal to the number of input reads.

Value

A GRanges object.

Examples

# get local address for included bam file
ps.bam <- system.file("extdata", "PROseq_dm6_chr4.bam",
                      package = "BRGenomics")

#--------------------------------------------------#
# Import entire reads
#--------------------------------------------------#

# Note that PRO-seq reads are sequenced as reverse complement
import_bam(ps.bam, revcomp = TRUE)

#> GRanges object with 53179 ranges and 1 metadata column:
#>           seqnames          ranges strand |     score
#>              <Rle>       <IRanges>  <Rle> | <integer>
#>       [1]     chr4       1189-1224      + |         1
#>       [2]     chr4     31442-31477      + |         1
#>       [3]     chr4     31707-31728      + |         1
#>       [4]     chr4     33805-33831      + |         1
#>       [5]     chr4     34253-34288      + |         1
#>       ...      ...             ...    ... .       ...
#>   [53175]     chr4 1307518-1307553      - |         1
#>   [53176]     chr4 1307615-1307650      - |         1
#>   [53177]     chr4 1316417-1316452      - |         1
#>   [53178]     chr4 1316506-1316541      - |         1
#>   [53179]     chr4 1316671-1316706      - |         2
#>   -------
#>   seqinfo: 1870 sequences from an unspecified genome

#--------------------------------------------------#
# Import entire reads, 1 range per read
#--------------------------------------------------#

import_bam(ps.bam, revcomp = TRUE, field = NULL)

#> GRanges object with 73887 ranges and 0 metadata columns:
#>           seqnames          ranges strand
#>              <Rle>       <IRanges>  <Rle>
#>       [1]     chr4       1189-1224      +
#>       [2]     chr4     31442-31477      +
#>       [3]     chr4     31707-31728      +
#>       [4]     chr4     33805-33831      +
#>       [5]     chr4     34253-34288      +
#>       ...      ...             ...    ...
#>   [73883]     chr4 1307615-1307650      -
#>   [73884]     chr4 1316417-1316452      -
#>   [73885]     chr4 1316506-1316541      -
#>   [73886]     chr4 1316671-1316706      -
#>   [73887]     chr4 1316671-1316706      -
#>   -------
#>   seqinfo: 1870 sequences from an unspecified genome

#--------------------------------------------------#
# Import PRO-seq reads at basepair-resolution
#--------------------------------------------------#

# the typical manner to import PRO-seq data:
import_bam(ps.bam, revcomp = TRUE, trim.to = "3p")

#> GRanges object with 48697 ranges and 1 metadata column:
#>           seqnames    ranges strand |     score
#>              <Rle> <IRanges>  <Rle> | <integer>
#>       [1]     chr4      1224      + |         1
#>       [2]     chr4     31477      + |         1
#>       [3]     chr4     31728      + |         1
#>       [4]     chr4     33831      + |         1
#>       [5]     chr4     34285      + |         1
#>       ...      ...       ...    ... .       ...
#>   [48693]     chr4   1307518      - |         1
#>   [48694]     chr4   1307615      - |         1
#>   [48695]     chr4   1316417      - |         1
#>   [48696]     chr4   1316506      - |         1
#>   [48697]     chr4   1316671      - |         2
#>   -------
#>   seqinfo: 1870 sequences from an unspecified genome

#--------------------------------------------------#
# Import 5' ends of PRO-seq reads
#--------------------------------------------------#

# will include bona fide TSSes as well as hydrolysis products
import_bam(ps.bam, revcomp = TRUE, trim.to = "5p")

#> GRanges object with 47380 ranges and 1 metadata column:
#>           seqnames    ranges strand |     score
#>              <Rle> <IRanges>  <Rle> | <integer>
#>       [1]     chr4      1189      + |         1
#>       [2]     chr4     31442      + |         1
#>       [3]     chr4     31707      + |         1
#>       [4]     chr4     33805      + |         1
#>       [5]     chr4     34253      + |         1
#>       ...      ...       ...    ... .       ...
#>   [47376]     chr4   1307553      - |         1
#>   [47377]     chr4   1307650      - |         1
#>   [47378]     chr4   1316452      - |         1
#>   [47379]     chr4   1316541      - |         1
#>   [47380]     chr4   1316706      - |         2
#>   -------
#>   seqinfo: 1870 sequences from an unspecified genome