Iterative Subsampling for Metaplotting

Iterative Subsampling for Metaplotting

md.meta.subsample(
  dataset.gr,
  regions.gr,
  binsize = 1,
  first.bin.start = 1,
  sample_name = deparse(substitute(dataset.gr)),
  n.iter = 1000,
  prop.subsample = 0.1,
  lower = 0.125,
  upper = 0.875,
  NF = 1,
  field = "score",
  remove.empty = FALSE,
  ncores = 1
)

Arguments

dataset.gr
A GRanges object in which signal is contained in metadata (typically in the "score" field).

regions.gr
A GRanges object containing intervals over which to metaplot. All ranges must have the same width.

binsize
The size of bin (number of columns, e.g. basepairs) to use for metaplotting. Especially important for metaplots over large/sparse regions.

first.bin.start
The relative start position of the first bin, e.g. if regions.gr begins at 50 bases upstream of the TSS, set first.bin.start = -50. This number only affects the x-values that are returned, which are provided as a convenience.

sample_name
Defaults to the name of dataset.gr.

n.iter
Number of random subsampling iterations to perform. Default is 1000.

prop.subsample
The proportion of the ranges in regions.gr (e.g. the proportion of genes) to subsample in each iteration. The default is 0.1.

lower
The lower quantile of subsampled signal means to return. The default is 0.125 (12.5th percentile).

upper
The upper quantile of subsampled signal means to return. The default is 0.875 (85.5th percentile).

NF
Optional normalization factor by which to multiply the counts.

field
The metadata field of dataset.gr to be counted.

remove.empty
A logical indicating whether regions without signal should be removed from the analysis.

ncores
Number of cores to use for parallel computation. As of writing, parallel processing doesn't show any benefit for short computation times (e.g. <1 minute for our typical experience on a laptop).

Value

Dataframe containing x-values, means, lower quantiles, upper quantiles, and the sample name (as a convenience for row-binding multiple of these dataframes).