Iterative Subsampling for Metaplotting (On Count Matrices)

In the most general sense, this function performs iterations of randomly subsampling rows of a matrix, and returns a summary of mean values calculated for each column. The typical application is for generating metaplots, with the typical input being a matrix in which each row is a gene or other region of interest, each column is a position within that gene (either a specific basepair or a bin), and element values are signal (e.g. read counts) within those positions.

metaSubsampleMatrix(
  counts.mat,
  binsize = 1,
  first.output.xval = 1,
  sample.name = deparse(substitute(counts.mat)),
  n.iter = 1000,
  prop.sample = 0.1,
  lower = 0.125,
  upper = 0.875,
  NF = 1,
  ncores = 1
)

Arguments

counts.mat
A matrix of signal counts in which rows are regions of interest and columns are sites/bins in each region.

binsize
The size of bin (number of columns, e.g. basepairs) to use for metaplotting. Especially important for metaplots over large/sparse regions.

first.output.xval
The relative start position of the first bin, e.g. if regions.gr begins at 50 bases upstream of the TSS, set first.output.xval = -50. This number only affects the x-values that are returned, which are provided as a convenience.

sample.name
Defaults to the name of dataset.gr.

n.iter
Number of random subsampling iterations to perform. Default is 1000.

prop.sample
The proportion of rows to subsample in each iteration. The default is 0.1.

lower
The lower quantile of subsampled signal means to return. The default is 0.125 (12.5th percentile).

upper
The upper quantile of subsampled signal means to return. The default is 0.875 (85.5th percentile).

NF
Optional normalization factor by which to multiply the counts.

ncores
Number of cores to use for parallel computation. As of writing, parallel processing doesn't show any benefit for short computation times (e.g. <1 minute for our typical experience on a laptop).

Value

Dataframe containing x-values, means, lower quantiles, upper quantiles, and the sample name (as a convenience for row-binding multiple of these dataframes).

See also

Examples

data("PROseq") # load included PROseq data
data("txs_dm6_chr4") # load included transcripts

# for each transcript, use promoter-proximal region from TSS to +100
pr <- promoters(txs_dm6_chr4, 0, 100)

# generate a matrix of counts in each region
countsmat <- getCountsByPositions(PROseq, pr)
dim(countsmat)

#> [1] 339 100

#--------------------------------------------------#
# bootstrap average signal in 10 bp bins across all transcripts
#--------------------------------------------------#

set.seed(11)
df <- metaSubsampleMatrix(countsmat, binsize = 10, sample.name = "PROseq")
df[1:10, ]

#>     x       mean     lower     upper sample.name
#> 1   6  0.4705882 0.2352941  1.147059      PROseq
#> 2  16  0.7058824 0.3529412  1.735294      PROseq
#> 3  26  1.2058824 0.7058824  2.176471      PROseq
#> 4  36  3.6470588 1.7647059 11.687500      PROseq
#> 5  46  5.6617647 2.9963235 11.709559      PROseq
#> 6  56 10.3382353 6.0000000 15.709559      PROseq
#> 7  66  5.2058824 3.2941176  7.411765      PROseq
#> 8  76  5.4558824 2.1433824 10.003676      PROseq
#> 9  86  3.0588235 2.0588235  4.441176      PROseq
#> 10 96  2.8676471 1.6470588  4.558824      PROseq

#--------------------------------------------------#
# the same, using a normalization factor, and changing the x-values
#--------------------------------------------------#

set.seed(11)
df <- metaSubsampleMatrix(countsmat, binsize = 10, first.output.xval = 0,
                          NF = 0.75, sample.name = "PROseq")
df[1:10, ]

#>     x      mean     lower      upper sample.name
#> 1   5 0.3529412 0.1764706  0.8602941      PROseq
#> 2  15 0.5294118 0.2647059  1.3014706      PROseq
#> 3  25 0.9044118 0.5294118  1.6323529      PROseq
#> 4  35 2.7352941 1.3235294  8.7656250      PROseq
#> 5  45 4.2463235 2.2472426  8.7821691      PROseq
#> 6  55 7.7536765 4.5000000 11.7821691      PROseq
#> 7  65 3.9044118 2.4705882  5.5588235      PROseq
#> 8  75 4.0919118 1.6075368  7.5027574      PROseq
#> 9  85 2.2941176 1.5441176  3.3308824      PROseq
#> 10 95 2.1507353 1.2352941  3.4191176      PROseq