Subset regions of interest by highest signal

Subsets regions based on signal in a dataset, taking only the top quantile of regions.

subsetRegionsBySignal(
  regions.gr,
  dataset.gr,
  regions_quantile,
  field = "score",
  order_by_rank = FALSE,
  density = FALSE
)

Arguments

regions.gr
A GRanges object containing regions of interest.

dataset.gr
A GRanges object in which signal is contained in metadata (typically in the "score" field).

regions_quantile
The proportion of regions.gr to return, e.g. if regions_quantile = 0.2, the top 20% of regions by signal are returned.

field
The metadata field of dataset.gr to be counted.

order_by_rank
Logical indicating if genes should be returned in order of their expression. If FALSE (the default), genes are sorted by their positions.

density
A logical indicating whether signal counts should be normalized to the width of ranges in regions.gr. By default, the function only considers the total signal in each range.

Value

A GRanges object of length length(regions.gr) * regions_quantile.