Extract Genebodies

This function returns ranges that are defined relative to the strand-specific start and end sites of regions of interest (usually genes). Unlike GenomicRanges::promoters, distances can be upstream or downstream based on the sign, and both the start and end of the returned regions can be defined in terms of either the start of end site of the input ranges. For example, genebodies(txs, -50, 150, fix.end = "start") is equivalent to promoters(txs, 50, 150). The default arguments return ranges that begin 300 bases downstream of the original start positions, and end 300 bases upstream of the original end positions.

genebodies(
  genelist,
  start = 300,
  end = -300,
  fix.start = "start",
  fix.end = "end",
  min.window = 0
)

Arguments

genelist
A GRanges object containing genes of interest.

start
Depending on fix.start, the distance from either the strand-specific start or end site to begin the returned ranges. If positive, the returned range will begin downstream of the reference position; negative numbers are used to return sites upstream of the reference. Set start = 0 to return the reference position.

end
Identical to the start argument, but defines the strand-specific end position of returned ranges. end must be downstream of start.

fix.start
The reference point to use for defining the strand-specific start positions of returned ranges, either "start" or "end".

fix.end
The reference point to use for defining the strand-specific end positions of returned ranges, either "start" or "end". Cannot be set to "start" if fix.start = "end".

min.window
When fix.start = "start" and fix.end = "end", min.window defines the minimum size (width) of a returned range. However, when fix.end = fix.start, all returned ranges have the same width, and min.window simply size-filters the input ranges.

Value

A GRanges object that may be shorter than genelist due to loss of short ranges.

See also

Examples

data("txs_dm6_chr4") # load included transcript data
len <- length(txs_dm6_chr4)
txs <- txs_dm6_chr4[c(1:2, len - 1, len)]
txs

#> GRanges object with 4 ranges and 2 metadata columns:
#>       seqnames          ranges strand |     tx_name     gene_id
#>          <Rle>       <IRanges>  <Rle> | <character> <character>
#>   [1]     chr4        879-5039      + | FBtr0346692 FBgn0267363
#>   [2]     chr4     42774-43374      + | FBtr0344900 FBgn0266617
#>   [3]     chr4 1225737-1230713      - | FBtr0100402 FBgn0027101
#>   [4]     chr4 1225737-1230713      - | FBtr0100404 FBgn0027101
#>   -------
#>   seqinfo: 7 sequences from dm6 genome

#--------------------------------------------------#
# genebodies from +300 (300 bp after TSS) to 300 bp before the poly-A site
#--------------------------------------------------#

genebodies(txs, 300, -300)

#> GRanges object with 4 ranges and 2 metadata columns:
#>       seqnames          ranges strand |     tx_name     gene_id
#>          <Rle>       <IRanges>  <Rle> | <character> <character>
#>   [1]     chr4       1179-4739      + | FBtr0346692 FBgn0267363
#>   [2]     chr4           43074      + | FBtr0344900 FBgn0266617
#>   [3]     chr4 1226037-1230413      - | FBtr0100402 FBgn0027101
#>   [4]     chr4 1226037-1230413      - | FBtr0100404 FBgn0027101
#>   -------
#>   seqinfo: 7 sequences from dm6 genome

#--------------------------------------------------#
# promoters from -50 to +100
#--------------------------------------------------#

promoters(txs, 50, 100)

#> GRanges object with 4 ranges and 2 metadata columns:
#>       seqnames          ranges strand |     tx_name     gene_id
#>          <Rle>       <IRanges>  <Rle> | <character> <character>
#>   [1]     chr4         829-978      + | FBtr0346692 FBgn0267363
#>   [2]     chr4     42724-42873      + | FBtr0344900 FBgn0266617
#>   [3]     chr4 1230614-1230763      - | FBtr0100402 FBgn0027101
#>   [4]     chr4 1230614-1230763      - | FBtr0100404 FBgn0027101
#>   -------
#>   seqinfo: 7 sequences from dm6 genome

genebodies(txs, -50, 100, fix.end = "start")

#> GRanges object with 4 ranges and 2 metadata columns:
#>       seqnames          ranges strand |     tx_name     gene_id
#>          <Rle>       <IRanges>  <Rle> | <character> <character>
#>   [1]     chr4         829-979      + | FBtr0346692 FBgn0267363
#>   [2]     chr4     42724-42874      + | FBtr0344900 FBgn0266617
#>   [3]     chr4 1230613-1230763      - | FBtr0100402 FBgn0027101
#>   [4]     chr4 1230613-1230763      - | FBtr0100404 FBgn0027101
#>   -------
#>   seqinfo: 7 sequences from dm6 genome

#--------------------------------------------------#
# region from 500 to 1000 bases after the poly-A site
#--------------------------------------------------#

genebodies(txs, 500, 1000, fix.start = "end")

#> GRanges object with 4 ranges and 2 metadata columns:
#>       seqnames          ranges strand |     tx_name     gene_id
#>          <Rle>       <IRanges>  <Rle> | <character> <character>
#>   [1]     chr4       5539-6039      + | FBtr0346692 FBgn0267363
#>   [2]     chr4     43874-44374      + | FBtr0344900 FBgn0266617
#>   [3]     chr4 1224737-1225237      - | FBtr0100402 FBgn0027101
#>   [4]     chr4 1224737-1225237      - | FBtr0100404 FBgn0027101
#>   -------
#>   seqinfo: 7 sequences from dm6 genome