Get DESeq2 results using reduced dispersion matrices

This function calls DESeq2::DESeq and DESeq2::results on a pre-existing DESeqDataSet object and returns a DESeqResults table for one or more pairwise comparisons. However, unlike a standard call to DESeq2::results using the contrast argument, this function subsets the dataset so that DESeq2 only estimates dispersion for the samples being compared, and not for all samples present.

getDESeqResults(
  dds,
  contrast.numer,
  contrast.denom,
  comparisons.list = NULL,
  sizeFactors = NULL,
  alpha = 0.1,
  args.DESeq = NULL,
  args.results = NULL,
  ncores = detectCores(),
  quiet = FALSE
)

Arguments

dds
A DESeqDataSet object, produced using either getDESeqDataSet from this package or DESeqDataSet from DESeq2. If dds was not created using getDESeqDataSet, dds must be made with design = ~condition such that a unique condition level exists for each sample/treatment condition.

contrast.numer
A string naming the condition to use as the numerator in the DESeq2 comparison, typically the perturbative condition.

contrast.denom
A string naming the condition to use as the denominator in the DESeq2 comparison, typically the control condition.

comparisons.list
As an optional alternative to supplying a single contrast.numer and contrast.denom, users can supply a list of character vectors containing numerator-denominator pairs, e.g. list(c("B", "A"), c("C", "A"), c("C", "B")).

sizeFactors
A vector containing DESeq2 sizeFactors to apply to each sample. Each sample's readcounts are divided by its respective DESeq2 sizeFactor. A warning will be generated if the DESeqDataSet already contains sizeFactors, and the previous sizeFactors will be over-written.

alpha
The significance threshold passed to DESeqResults. This won't affect the output results, but is used as a performance optimization by DESeq2.

args.DESeq
Additional arguments passed to DESeq, given as a list of argument-value pairs, e.g. list(test = "LRT", fitType = "local"). All arguments given here will be passed to DESeq except for object and parallel. If no arguments are given, all defaults will be used.

args.results
Additional arguments passed to DESeq2::results, given as a list of argument-value pairs, e.g. list(altHypothesis = "greater", lfcThreshold = 1.5). All arguments given here will be passed to results except for object, contrast, alpha, and parallel. If no arguments are given, all defaults will be used.

ncores
The number of cores to use for parallel processing. Multicore processing is only used if more than one comparison is being made (i.e. argument comparisons.list is used), and the number of cores utilized will not be greater than the number of comparisons being performed.

quiet
If TRUE, all output messages from calls to DESeq and results will be suppressed, although passing option quiet in args.DESeq will supersede this option for the call to DESeq.

Value

For a single comparison, the output is the DESeqResults result table. If a comparisons.list is used to make multiple comparisons, the output is a named list of DESeqResults objects, with elements named following the pattern "X_vs_Y", where X is the name of the numerator condition, and Y is the name of the denominator condition.

See also

Examples

#--------------------------------------------------#
# getDESeqDataSet
#--------------------------------------------------#
suppressPackageStartupMessages(require(DESeq2))
data("PROseq") # import included PROseq data
data("txs_dm6_chr4") # import included transcripts

# divide PROseq data into 6 toy datasets
ps_a_rep1 <- PROseq[seq(1, length(PROseq), 6)]
ps_b_rep1 <- PROseq[seq(2, length(PROseq), 6)]
ps_c_rep1 <- PROseq[seq(3, length(PROseq), 6)]

ps_a_rep2 <- PROseq[seq(4, length(PROseq), 6)]
ps_b_rep2 <- PROseq[seq(5, length(PROseq), 6)]
ps_c_rep2 <- PROseq[seq(6, length(PROseq), 6)]

ps_list <- list(A_rep1 = ps_a_rep1,
                A_rep2 = ps_a_rep2,
                B_rep1 = ps_b_rep1,
                B_rep2 = ps_b_rep2,
                C_rep1 = ps_c_rep1,
                C_rep2 = ps_c_rep2)

# make flawed dataset (ranges in txs_dm6_chr4 not disjoint)
#    this means there is double-counting
# also using discontinuous gene regions, as gene_ids are repeated
dds <- getDESeqDataSet(ps_list,
                       txs_dm6_chr4,
                       gene_names = txs_dm6_chr4$gene_id,
                       ncores = 2)

#> renaming the first element in assays to 'counts'

#> converting counts to integer mode

dds

#> class: DESeqDataSet 
#> dim: 111 6 
#> metadata(1): version
#> assays(1): counts
#> rownames(111): FBgn0267363 FBgn0266617 ... FBgn0039924 FBgn0027101
#> rowData names(2): tx_name gene_id
#> colnames(6): A_rep1 A_rep2 ... C_rep1 C_rep2
#> colData names(2): condition replicate

#--------------------------------------------------#
# getDESeqResults
#--------------------------------------------------#

res <- getDESeqResults(dds, "B", "A")

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res

#> log2 fold change (MLE): condition B vs A 
#> Wald test p-value: condition B vs A 
#> DataFrame with 111 rows and 6 columns
#>               baseMean log2FoldChange     lfcSE        stat    pvalue      padj
#>              <numeric>      <numeric> <numeric>   <numeric> <numeric> <numeric>
#> FBgn0267363    0.25000     -1.4283946  4.997107 -0.28584433  0.774997   0.99766
#> FBgn0266617   10.34365      0.0838289  0.897865  0.09336470  0.925614   0.99766
#> FBgn0265633    2.01884      0.0120837  1.785163  0.00676898  0.994599   0.99766
#> FBgn0264617   64.61865      0.2152746  0.390560  0.55119533  0.581500   0.99766
#> FBgn0085432 4575.85639     -0.2116051  0.207035 -1.02207177  0.306747   0.99766
#> ...                ...            ...       ...         ...       ...       ...
#> FBgn0039928    696.376     0.00585552  0.242563   0.0241402  0.980741   0.99766
#> FBgn0052017    121.156    -0.04873622  0.338569  -0.1439476  0.885542   0.99766
#> FBgn0002521    172.065    -0.14843126  0.307854  -0.4821480  0.629701   0.99766
#> FBgn0039924    162.633     0.30042713  0.465331   0.6456196  0.518526   0.99766
#> FBgn0027101    357.788    -0.12638095  0.236357  -0.5347032  0.592855   0.99766

reslist <- getDESeqResults(dds,
                           comparisons.list = list(c("B", "A"), c("C", "A")),
                           ncores = 1)
names(reslist)

#> [1] "B_vs_A" "C_vs_A"

reslist[[1]]

#> log2 fold change (MLE): condition B vs A 
#> Wald test p-value: condition B vs A 
#> DataFrame with 111 rows and 6 columns
#>               baseMean log2FoldChange     lfcSE        stat    pvalue      padj
#>              <numeric>      <numeric> <numeric>   <numeric> <numeric> <numeric>
#> FBgn0267363    0.25000     -1.4283946  4.997107 -0.28584433  0.774997   0.99766
#> FBgn0266617   10.34365      0.0838289  0.897865  0.09336470  0.925614   0.99766
#> FBgn0265633    2.01884      0.0120837  1.785163  0.00676898  0.994599   0.99766
#> FBgn0264617   64.61865      0.2152746  0.390560  0.55119533  0.581500   0.99766
#> FBgn0085432 4575.85639     -0.2116051  0.207035 -1.02207177  0.306747   0.99766
#> ...                ...            ...       ...         ...       ...       ...
#> FBgn0039928    696.376     0.00585552  0.242563   0.0241402  0.980741   0.99766
#> FBgn0052017    121.156    -0.04873622  0.338569  -0.1439476  0.885542   0.99766
#> FBgn0002521    172.065    -0.14843126  0.307854  -0.4821480  0.629701   0.99766
#> FBgn0039924    162.633     0.30042713  0.465331   0.6456196  0.518526   0.99766
#> FBgn0027101    357.788    -0.12638095  0.236357  -0.5347032  0.592855   0.99766